third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-composite-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-composite with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-composite">
<meta name="assert" content="mask-composite supports only the grammar '<compositing-operator>#' where <compositing-operator> is 'add | subtract | intersect | exclude'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-composite", "none");
test_invalid_value("mask-composite", "exclude intersect subtract add");
</script>
</body>
</html>
